pyproject.toml
[project]
name = "political-polarization-chat"
version = "0.1.0"
description = "MLDS Final Project - Conversational Agents for Reducing Political Polarization"
authors = [
    {name = "Niv Bar", email = "niv.data.science@gmail.com"}  # Update with your real email
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core Streamlit and AI
    "streamlit>=1.48.0,<2.0.0",
    "google-genai>=1.29.0,<2.0.0",
    "google-generativeai>=0.8.5,<0.9.0",

    # Configuration and environment
    "python-dotenv>=1.0.0,<2.0.0",

    # Database - using Firebase
    "firebase-admin>=6.0.0,<7.0.0",

    # Basic utilities
    "requests>=2.31.0,<3.0.0",
]

# Development dependencies (optional)
[project.optional-dependencies]
dev = [
    "notebook>=7.4.5,<8.0.0",
    "jupyter>=1.1.1,<2.0.0",
    "ipywidgets>=8.1.7,<9.0.0",
    "pytest>=7.4.0,<8.0.0",
    "black>=23.0.0,<24.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"